% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.penalized-package}
\alias{mlr3learners.penalized}
\alias{mlr3learners.penalized-package}
\title{mlr3learners.penalized: Learners from the {penalized} package for 'mlr3'}
\description{
Adds learners from the {penalized} package to 'mlr3'.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk}

}
